Refactoring du code, extraction des méthodes du Main
