+++ src/main/Commande.java
@@ -1,4 +1,4 @@
-package partie2;
+package main;
 
 import java.util.ArrayList;
