+++ src/test/TestMain.java
@@ -1,0 +1,14 @@
+package test;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class TestMain {
+
+	@Test
+	public void test() {
+		fail("Not yet implemented");
+	}
+
+}
